+++ .swcrc
@@ -10,7 +10,7 @@
         "topLevel": false
       }
     },
-    "target": "es2021",
+    "target": "esnext",
     "parser": {
       "syntax": "typescript",
       "tsx": false,
