+++ src/engine/impl/index.ts
@@ -2,7 +2,6 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from "./FileEngine.js";
 export * from "./av/ClamAvEngine.js";
 export * from "./av/MsDefenderEngine.js";
 export * from "./HttpErrorRenderers/AuthenticationErrorRenderEngine.js";
