+++ src/public/assets/custom/css/index.css
@@ -117,6 +117,6 @@
     font-size: 2rem;
 }
 
-.am5-modal-content{
+.am5-modal-content {
     color: #000000 !important;
 }
