+++ src/public/index.ejs
@@ -12,7 +12,7 @@
             <a href="/" class="text-body-emphasis text-decoration-none">
                 <span class="fs-4">WaifuVault</span>
             </a>
-            <a class="btn btn-secondary" href="https://github.com/VictoriqueMoe/WaifuVault" target="_blank"><i class="bi bi-github"></i> Source</a>
+            <a class="btn btn-secondary" href="https://github.com/waifuvault/WaifuVault" target="_blank"><i class="bi bi-github"></i> Source</a>
         </header>
 
         <div class="p-5 mb-4 bg-body-tertiary rounded-3 border border-primary-subtle position-relative">
@@ -22,7 +22,7 @@
                     <a href="https://0x0.st" target="_blank">https://0x0.st</a>, WaifuVault is a temporary file hosting service that allows for file uploads that are hosted for a set amount of time.
                 </p>
                 <a class="btn btn-primary btn-lg" type="button" href="/api-docs" target="_blank">Api Documentation</a>
-                <img id="victorique" class="img-fluid" src="assets/custom/images/08.png" alt="logo"/>
+                <img id="victorique" class="img-fluid" src="assets/custom/images/08.png" alt="logo" />
             </div>
         </div>
